requirements-dev.txt
anyascii==0.2.0
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.44.0
django-cogwheels==0.3
django-extensions==3.1.3
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
entrypoints==0.3
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
invoke==1.5.0
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
l18n==2020.6.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
oauthlib==3.1.1
openpyxl==3.0.7
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
python-environ==0.4.54
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
Send2Trash==1.7.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.2
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wagtail==2.13.3
wagtailmenus==3.0.2
wcwidth==0.2.5
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.4
xlwt==1.3.0
